change price datatype to Double

+++ src/entity/Sale.java
@@ -6,7 +6,6 @@
 package entity;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
@@ -46,7 +45,7 @@ public class Sale implements Serializable
     private Integer saleQty;
     // @Max(value=?)  @Min(value=?)//if you know range of your decimal fields consider using these annotations to enforce field validation
     @Column(name = "sale_price")
-    private BigDecimal salePrice;
+    private Double salePrice;
     @Column(name = "sale_date")
     @Temporal(TemporalType.DATE)
     private Date saleDate;
@@ -88,12 +87,12 @@ public class Sale implements Serializable
         this.saleQty = saleQty;
     }
 
-    public BigDecimal getSalePrice()
+    public Double getSalePrice()
     {
         return salePrice;
     }
 
-    public void setSalePrice(BigDecimal salePrice)
+    public void setSalePrice(Double salePrice)
     {
         this.salePrice = salePrice;
     }
